requirements.txt
streamlit==1.30.0
pymupdf==1.23.1
pytesseract==0.3.10
Pillow==10.0.0
camelot-py[cv]==0.10.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.40.0
torch==2.1.0
